fix: type error when saving parquet

+++ config/schemas.py
@@ -2,16 +2,16 @@ from pyspark.sql import SparkSession, DataFrame
 from pyspark.sql.types import *
 
 yellow_taxi_schema = StructType([
-    StructField("VendorID", IntegerType(), True),
+    StructField("VendorID", LongType(), True),
     StructField("tpep_pickup_datetime", TimestampType(), True),
     StructField("tpep_dropoff_datetime", TimestampType(), True),
-    StructField("passenger_count", IntegerType(), True),
+    StructField("passenger_count", LongType(), True),
     StructField("trip_distance", DoubleType(), True),
-    StructField("RatecodeID", IntegerType(), True),
+    StructField("RatecodeID", LongType(), True),
     StructField("store_and_fwd_flag", StringType(), True),
-    StructField("PULocationID", IntegerType(), True),
-    StructField("DOLocationID", IntegerType(), True),
-    StructField("payment_type", IntegerType(), True),
+    StructField("PULocationID", LongType(), True),
+    StructField("DOLocationID", LongType(), True),
+    StructField("payment_type", LongType(), True),
     StructField("fare_amount", DoubleType(), True),
     StructField("extra", DoubleType(), True),
     StructField("mta_tax", DoubleType(), True),
@@ -25,14 +25,14 @@ yellow_taxi_schema = StructType([
 ])
 
 green_taxi_schema = StructType([
-    StructField("VendorID", IntegerType(), True),
+    StructField("VendorID", LongType(), True),
     StructField("lpep_pickup_datetime", TimestampType(), True),
     StructField("lpep_dropoff_datetime", TimestampType(), True),
     StructField("store_and_fwd_flag", StringType(), True),
-    StructField("RatecodeID", IntegerType(), True),
-    StructField("PULocationID", IntegerType(), True),
-    StructField("DOLocationID", IntegerType(), True),
-    StructField("passenger_count", IntegerType(), True),
+    StructField("RatecodeID", LongType(), True),
+    StructField("PULocationID", LongType(), True),
+    StructField("DOLocationID", LongType(), True),
+    StructField("passenger_count", LongType(), True),
     StructField("trip_distance", DoubleType(), True),
     StructField("fare_amount", DoubleType(), True),
     StructField("extra", DoubleType(), True),
@@ -42,8 +42,8 @@ green_taxi_schema = StructType([
     StructField("ehail_fee", DoubleType(), True),
     StructField("improvement_surcharge", DoubleType(), True),
     StructField("total_amount", DoubleType(), True),
-    StructField("payment_type", IntegerType(), True),
-    StructField("trip_type", IntegerType(), True),
+    StructField("payment_type", LongType(), True),
+    StructField("trip_type", LongType(), True),
     StructField("congestion_surcharge", DoubleType(), True),
     StructField("cbd_congestion_fee", DoubleType(), True),
 ])
